Fix error when duplicate filters are generated.
This is necessary because Gmail reports an error when trying to
create a filter that is already present. The error reported is
quite uninformative, so it's just better to get rid of those
transparently.

The implementation of this fix is not really great, but it's the
easiest at this point without refactoring. The diff command can
already detect duplicates; now it also makes sure they are not
reported.