Reverse "want" and "got" in test failures
Suppose we have a rule configured to label messages with
"oops-bad-label", and a test that checks they're labeled with
"yep-correct-label". So before this change you would get:

    --- want
    +++ got
    ...
      "labels": [
    -   "oops-bad-label"
    +   "yep-correct-label"
      ]

To make this match the usual logic of automated testing (especially Go's
"want" and "got"), we would like the "want" section to be "what the test
describes should happen". Therefore, with this change, we get this
output:

    --- want
    +++ got
    ...
      "labels": [
    -   "yep-correct-label"
    +   "oops-bad-label"
      ]